Dict: enable faster hashing for homogeneous keys
Hashing is based on "`isequal` implies hashes equal". So for example as we have `1.0 == 1`,
then `hash(1.0) == hash(1)`. The point of this rule is to have dictionaries (and sets)
behave as we have come to expect, i.e. `d = Dict(1=>2); @assert d[1.0] == 2`.

But this rule comes at a cost. Probably one of the most spectacular examples is for
ranges, e.g `hash(1:2^60)` takes half a second!
This is so that this hash is equal to `hash(collect(1:2^60))`. But, what if
my dictionary contains only unit ranges? Then a more specific/fast hashing algorithm
can be chosen.

This PR introduces an API for 3-argument hash methods, `hash(::Type{K}, x::K, h::UInt)`.
The first argument `K` specifies the contextual type, i.e. for a `Dict{K}` and
a key `x`, `hash(K, x, h)` will be called (it defaults of course to `hash(x, h)`).

Currently, methods are written for `Integer`, `AbstractUnitRange` and `AbstractSet`.
Here is a little demo showing timings: